check to see if they have a custom avatar that needs removing before trying to remove it (should prevent error from showing, will remove the ugly message if it tests ok)
git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@38865 e846bd91-c042-42a7-a218-4b65e2c1904d

+++ extensions/wikia/EditAccount/SpecialEditAccount_body.php
@@ -256,11 +256,12 @@ class EditAccount extends SpecialPage {
 		// remove users avatar
 		if ( class_exists( Masthead ) ) {
 			$avatar = Masthead::newFromUser( $this->mUser );
-			if ( !$avatar->removeFile( false ) ) {
-				// $this->mStatusMsg = wfMsg( 'editaccount-error-close', $this->mUser->mName );
-				// dont quit here, since the avatar is a non-critical part of closing.
-				// temp fix: move this to i18n message after looking into ->removeFile
-				$outMsg = "Problem removing avatar (or no avatar to remove) Use dedicated tool.<br/>\n";
+			if ( !$avatar->isDefault() ) {
+				if( !$avatar->removeFile( false ) ) {
+					# dont quit here, since the avatar is a non-critical part of closing.
+					# temp fix: move this to i18n message after looking into ->removeFile
+					$outMsg = "Problem removing avatar. Use dedicated tool.\n";
+				}
 			}
 		}
